move auction cache to adapters and nest inside generic cache
