+++ packages/router/src/lib/entities/cache.ts
@@ -1,0 +1,9 @@
+import { BigNumber } from "ethers";
+
+export type CachedBid = {
+  assetId: string;
+  chainId: number;
+  amountReceived: BigNumber;
+  expiry: number;
+  transactionId: string;
+};
